fix: display milliseconds, matching given labels (#127)
The [ActiveSupport::Notifications::Event class][1] states that times are
stored in milliseconds. Based on the label `ms` in [badge.html.erb][2]
and [profiled_requests/show.html.erb][github], I would expect the
numbers show next to those labels to also be milliseconds. However, what
is actually stored is microseconds, calculated by multiplying the
`Event` millisecond values by 100,000 in [tracer.rb][github 2].  As a
result, the numbers shown next to the label `ms` are actually µs
instead. To fix this, we need to divide the integer values by 100,000
again. That's what this PR does.

[1]: https://github.com/rails/rails/blob/76489d81ba77216271870e11fba6889088016fa5/activesupport/lib/active_support/notifications/instrumenter.rb#L120-L134
[2]: https://github.com/hschne/rails-mini-profiler/blob/main/app/views/rails_mini_profiler/badge.html.erb#L36
[github]: https://github.com/hschne/rails-mini-profiler/blob/main/app/views/rails_mini_profiler/profiled_requests/show.html.erb#L17
[github 2]: https://github.com/hschne/rails-mini-profiler/blob/main/lib/rails_mini_profiler/tracers/tracer.rb#L31-L32